Refactor: dndkitの操作による不要なレンダリングを防ぐ
